Add CI
To make this work required removing the base tag from the index.html
file, which then forces the app to use relative URLs. The
serviceworker hack from ThunderstoneQuest Randomizer is not necessary
any more either, it seems. However, to run speech_to_text in Web
requires using the non-stable version of that library, so that's been
updated in the pubspec.